+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.6.0] - 2026-07-17
 
 ### Added
 
@@ -99,6 +99,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial stable version.
 
+[1.6.0]: https://github.com/ozontech/testo/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/ozontech/testo/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/ozontech/testo/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/ozontech/testo/compare/v1.3.2...v1.4.0